+++ lib/td_dd/audit/audit_support.ex
@@ -5,6 +5,7 @@ defmodule TdDd.Audit.AuditSupport do
 
   alias Ecto.Changeset
   alias TdCache.Audit
+  alias TdDd.DataStructures.StructureNote
   alias TdDfLib.{MapDiff, Masks}
 
   def publish(event, resource_type, resource_id, user_id, payload \\ %{})
@@ -31,6 +32,11 @@ defmodule TdDd.Audit.AuditSupport do
       user_id: user_id,
       payload: payload
     )
+  end
+
+  defp payload(%{df_content: new_content}, %StructureNote{df_content: old_content} = _data) do
+    diff = MapDiff.diff(old_content, new_content, mask: &Masks.mask/1)
+    %{content: diff}
   end
 
   defp payload(%{df_content: new_content} = changes, %{df_content: old_content} = _data) do