+++ lib/td_dd/search/mappings.ex
@@ -51,6 +51,7 @@ defmodule TdDd.Search.Mappings do
       domain_ids: %{type: "long"},
       deleted_at: %{type: "date", format: "strict_date_optional_time||epoch_millis"},
       last_change_by: %{enabled: false},
+      metadata: %{enabled: false},
       linked_concepts_count: %{type: "short"},
       inserted_at: %{type: "date", format: "strict_date_optional_time||epoch_millis"},
       updated_at: %{type: "date", format: "strict_date_optional_time||epoch_millis"},