+++ lib/td_dq_web/views/implementation_view.ex
@@ -29,7 +29,8 @@ defmodule TdDqWeb.ImplementationView do
       :execution_result_info,
       :structure_aliases,
       :df_name,
-      :df_content
+      :df_content,
+      :executable
     ])
     |> Map.put(
       :raw_content,
@@ -52,7 +53,8 @@ defmodule TdDqWeb.ImplementationView do
       :execution_result_info,
       :structure_aliases,
       :df_name,
-      :df_content
+      :df_content,
+      :executable
     ])
     |> Map.put(:dataset, render_many(implementation.dataset, DatasetView, "dataset_row.json"))
     |> Map.put(