+++ CHANGELOG.md
@@ -2,10 +2,17 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- [TD-3893] Children classifications
+- [TD-3905] Fix bug with StructureNote aggregation
+- [TD-3907] Fix metadata index failure
+
 ### Added
 
 - [TD-3720] Update structure domain (with children)
 - [TD-3522] Support for StructureNote management with workflow
+- [TD-3552] Executable implementations
 
 ## [4.22.0] 2021-06-15